+++ packages/syntax-core/src/Badge/Badge.tsx
@@ -1,7 +1,9 @@
 import Typography from "../Typography/Typography";
 import Box from "../Box/Box";
+import styles from "./Badge.module.css";
 
 const BadgeColor = [
+  "gray200",
   "gray900",
   "destructive700",
   "orange700",
@@ -11,13 +13,30 @@ const BadgeColor = [
   "purple700",
 ] as const;
 
+const textColorForBackgroundColor = (
+  color: (typeof BadgeColor)[number],
+): "gray900" | "white" => {
+  switch (color) {
+    case "gray200":
+    case "yellow700":
+      return "gray900";
+    default:
+      return "white";
+  }
+};
+
 /**
- * Badge is a component to display short text and give additional context to features and other components.
+ * [Badge](https://cambly-syntax.vercel.app/?path=/docs/components-badge--docs) is a component to display short text and give additional context to features and other components.
  */
 const Badge = ({
+  icon: Icon,
   text,
   color = "primary700",
 }: {
+  /**
+   * The icon to be displayed. Please use a [Material Icon](https://material.io/resources/icons/)
+   */
+  icon?: React.ComponentType<{ className: string }>;
   /**
    * The text to display inside the badge
    */
@@ -35,13 +54,19 @@ const Badge = ({
     paddingY={1}
     rounding="full"
     backgroundColor={color}
+    dangerouslySetInlineStyle={{ __style: { lineHeight: "14px" } }}
   >
-    <Typography
-      color={color === "yellow700" ? "gray900" : "white"}
-      size={100}
-      weight="bold"
-    >
-      {text}
+    <Typography color={textColorForBackgroundColor(color)} size={100}>
+      <Box display="flex" gap={1} alignItems="center" justifyContent="start">
+        {Icon && <Icon className={styles.icon} />}
+        <Typography
+          color={textColorForBackgroundColor(color)}
+          size={100}
+          weight="bold"
+        >
+          {text}
+        </Typography>
+      </Box>
     </Typography>
   </Box>
 );