+++ packages/syntax-core/src/Button/Button.test.tsx
@@ -42,6 +42,21 @@ describe("button", () => {
     expect(button).toHaveLength(1);
   });
 
+  it("correctly applies fullWidth when set", async () => {
+    render(
+      <Button
+        onClick={() => {
+          /* empty */
+        }}
+        text="Continue"
+        data-testid="button-test-id"
+        fullWidth
+      />,
+    );
+    const button = await screen.findByTestId("button-test-id");
+    expect(button).toHaveStyle({ width: "100%" });
+  });
+
   it("sets an accessibility label", async () => {
     render(
       <Button
@@ -66,9 +81,24 @@ describe("button", () => {
       />,
     );
     const button = await screen.findAllByLabelText("Continue to the next step");
-    // eslint-disable-next-line testing-library/no-await-sync-events
     await userEvent.click(button[0]);
     expect(handleClick).toHaveBeenCalledTimes(1);
+  });
+
+  it("fires the onSubmit event on a form when button of type submit is clicked", async () => {
+    const handleSubmit = vi.fn();
+    render(
+      <form onSubmit={handleSubmit}>
+        <Button
+          type="submit"
+          text="Continue"
+          accessibilityLabel="Continue to the next step"
+        />
+      </form>,
+    );
+    const button = await screen.findByLabelText("Continue to the next step");
+    await userEvent.click(button);
+    expect(handleSubmit).toHaveBeenCalledTimes(1);
   });
 
   it("sets the data-testid", () => {