+++ packages/syntax-core/src/Button/constants/iconSize.ts
@@ -1,0 +1,9 @@
+import styles from "../Button.module.css";
+
+const iconSize = {
+  ["sm"]: styles.smIcon,
+  ["md"]: styles.mdIcon,
+  ["lg"]: styles.lgIcon,
+};
+
+export default iconSize;