+++ packages/syntax-core/src/MiniActionCard/MiniActionCard.tsx
@@ -1,7 +1,7 @@
 import styles from "./MiniActionCard.module.css";
 
 /**
- * MiniActionCard is component that alerts users to a call to action.
+ * [MiniActionCard](https://cambly-syntax.vercel.app/?path=/docs/components-miniactioncard--docs) is component that alerts users to a call to action.
  */
 const MiniActionCard = ({
   children,