+++ packages/syntax-floating-components/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @cambly/syntax-floating-components
 
+## 0.2.0
+
+### Minor Changes
+
+- 8be2304: Button/RadioButton/Checkbox/SelectList/IconButton: update disabled opacity
+
+## 0.1.4
+
+### Patch Changes
+
+- ad32c4b: Add individual links to components
+
+## 0.1.3
+
+### Patch Changes
+
+- f1f5b2f: Fix tests
+
 ## 0.1.2
 
 ### Patch Changes